Rename rpc_electrum.rs to rpc_client.rs, swap_ops.rs to swap.rs
